Add numerical method realization

+++ src/main/java/computations/functions/FunctionContainer.java
@@ -3,6 +3,7 @@ package computations.functions;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.function.Function;
 
 @AllArgsConstructor
@@ -11,6 +12,6 @@ public abstract class FunctionContainer {
     @Getter
     private String description;
 
-    public abstract Function<Double, Double> getFunction();
+    public abstract Function<BigDecimal, BigDecimal> getFunction();
 
 }
